Track final CNB research bundle
Remove obsolete input manifests, retain the verified EUCF output bundle, split CNB research tests by responsibility, and keep the MLflow experiment name lowercase.
