+++ climate-advisor/.env.example
@@ -12,7 +12,7 @@ CA_FEATURE_FLAGS=
 MLFLOW_ENABLED=true
 MLFLOW_TRACKING_URI=https://mlflow-dev.openearth.dev
 MLFLOW_ENVIRONMENT=dev
-MLFLOW_EXPERIMENT_NAME=Clima
+MLFLOW_EXPERIMENT_NAME=clima
 MLFLOW_HTTP_REQUEST_TIMEOUT=3
 MLFLOW_HTTP_REQUEST_MAX_RETRIES=1
 MLFLOW_HTTP_REQUEST_BACKOFF_FACTOR=1
