--- climate-advisor/scripts/cnb_research/files/european_city_facility_call_7.json
@@ -1,8 +1,0 @@
-{
-  "funder_name": "European Union LIFE Programme",
-  "funder_url": "https://cinea.ec.europa.eu/programmes/life_en",
-  "program_name": "European City Facility (EUCF) - Call 7",
-  "program_url": "https://www.eucityfacility.eu/calls/call-7",
-  "application_template_url": null,
-  "max_turns": 12
-}
