--- climate-advisor/scripts/cnb_research/files/local_climate_action_grants.json
@@ -1,8 +1,0 @@
-{
-  "funder_name": "Minnesota Pollution Control Agency",
-  "funder_url": "https://www.pca.state.mn.us/",
-  "program_name": "Local Climate Action Grants",
-  "program_url": "https://www.pca.state.mn.us/grants-and-loans/local-climate-action-grants",
-  "application_template_url": null,
-  "max_turns": 10
-}
