--- climate-advisor/scripts/cnb_research/files/solar_on_public_buildings.json
@@ -1,8 +1,0 @@
-{
-  "funder_name": "Minnesota Department of Commerce",
-  "funder_url": "https://mn.gov/commerce/",
-  "program_name": "Solar on Public Buildings Grant Program",
-  "program_url": "https://mn.gov/commerce/energy/consumer/energy-programs/solar-public-buildings.jsp",
-  "application_template_url": null,
-  "max_turns": 10
-}
