+++ climate-advisor/service/tests/test_mlflow_logging.py
@@ -122,7 +122,7 @@ def test_mlflow_experiment_name_matches_active_server_name(monkeypatch) -> None:
     """The default preserves the case-sensitive active MLflow experiment name."""
     monkeypatch.delenv("MLFLOW_EXPERIMENT_NAME", raising=False)
 
-    assert mlflow_logging.climate_advisor_experiment_name() == "Clima"
+    assert mlflow_logging.climate_advisor_experiment_name() == "clima"
 
 
 def test_live_span_set_tag_compatibility_uses_span_attributes(monkeypatch) -> None:
